requirements.txt
django==4.2.14
pillow
gunicorn
python-dotenv
python-dateutil